refactor: creating src repository with App.js as main entry point
